move getMobileOperatingSystem to utils
