+++ express/code/blocks/app-banner/app-banner.js
@@ -1,7 +1,6 @@
-import { getLibs } from '../../scripts/utils.js';
+import { getLibs, getMobileOperatingSystem } from '../../scripts/utils.js';
 import { fixIcons, getIconElementDeprecated } from '../../scripts/utils/icons.js';
 import { addTempWrapperDeprecated, decorateButtonsDeprecated } from '../../scripts/utils/decorate.js';
-import { getMobileOperatingSystem } from '../../scripts/utils/media.js';
 
 let createTag; let getConfig;
 let replaceKeyArray; let appleStoreRatingScore;
