+++ express/code/blocks/collapsible-card/collapsible-card.js
@@ -1,8 +1,7 @@
-import { getLibs } from '../../scripts/utils.js';
+import { getLibs, getMobileOperatingSystem } from '../../scripts/utils.js';
 import { addTempWrapperDeprecated } from '../../scripts/utils/decorate.js';
 import { getIconElementDeprecated } from '../../scripts/utils/icons.js';
 import { fetchRelevantRows } from '../../scripts/utils/relevant.js';
-import { getMobileOperatingSystem } from '../../scripts/utils/media.js';
 
 let createTag;
