+++ express/code/scripts/utils.js
@@ -48,6 +48,25 @@ export function getCachedMetadata(name) {
   return cachedMetadata[name];
 }
 
+export function getMobileOperatingSystem() {
+  const userAgent = navigator.userAgent || navigator.vendor || window.opera;
+
+  // Windows Phone must come first because its UA also contains "Android"
+  if (/windows phone/i.test(userAgent)) {
+    return 'Windows Phone';
+  }
+
+  if (/android/i.test(userAgent)) {
+    return 'Android';
+  }
+
+  if (/iPad|iPhone|iPod/.test(userAgent) && !window.MSStream) {
+    return 'iOS';
+  }
+
+  return 'unknown';
+}
+
 export async function getRedirectUri() {
   const { getConfig } = await import(`${getLibs()}/utils/utils.js`);
   if (getMetadata('adobe-home-redirect') === 'on') {
