+++ express/code/scripts/widgets/floating-cta.js
@@ -1,9 +1,8 @@
 // import { decorateLinks} from '../';
 
-import { getLibs, getLottie, lazyLoadLottiePlayer, createTag } from '../utils.js';
+import { getLibs, getLottie, lazyLoadLottiePlayer, createTag, getMobileOperatingSystem } from '../utils.js';
 import { getIconElementDeprecated } from '../utils/icons.js';
 import BlockMediator from '../block-mediator.min.js';
-import { getMobileOperatingSystem } from '../utils/media.js';
 
 export const hideScrollArrow = (floatButtonWrapper, lottieScrollButton) => {
   floatButtonWrapper.classList.add('floating-button--scrolled');
